refactor: organize components folder structure
Move components around and rename them in the components folder.
