+++ src/app/components/Common/LocationImage/LocationImage.tsx
