+++ src/app/components/Common/LocationImage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LocationImage';
