+++ src/app/components/Common/Map/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Map';
