+++ src/app/components/Features/Game/GameContainer/GameContainer.css
