+++ src/app/components/Features/Game/GameContainer/GameContainer.tsx
@@ -1,9 +1,9 @@
 'use client';
 import React, { useState } from 'react';
-import GameLocationImage from '@components/Features/Locations/GameLocationImage';
+import GameLocationImage from '@components/Features/Game/GameLocationImage';
 import MapType from '@typings/MapType';
 import dynamic from 'next/dynamic';
-const Map = dynamic(() => import('@components/Features/Map'), {
+const GameMap = dynamic(() => import('@components/Features/Game/GameMap'), {
     loading: () => <p>Map is loading...</p>,
     ssr: false,
 });
@@ -25,7 +25,7 @@ const GameContainer: React.FC = () => {
                     // console.log('set true');
                 }}
             >
-                <Map
+                <GameMap
                     mapType={MapType.SFUBurnaby}
                     className="w-full h-full z-0"
                     minimized={minimized}
