+++ src/app/components/Features/Game/GameContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GameContainer';
