+++ src/app/components/Features/Game/GameLocationImage/GameLocationImage.css
