+++ src/app/components/Features/Game/GameLocationImage/GameLocationImage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import GameLocationImageProps from '@typings/GameLocationImageProps';
-import LocationImage from './LocationImage';
+import LocationImage from '@components/Common/LocationImage/LocationImage';
 import './GameLocationImage.css';
 
 const GameLocationImage: React.FC<GameLocationImageProps> = (
