+++ src/app/components/Features/Game/GameLocationImage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GameLocationImage';
