+++ src/app/components/Features/Game/GameMap/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GameMap';
