+++ src/app/components/Layouts/Navbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Navbar';
